apps/dashboard: make the create for bplans and external projects work
