+++ meinberlin/apps/extprojects/templates/meinberlin_extprojects/external_project_create_form.html
@@ -1,12 +1,12 @@
 {% extends "a4dashboard/base_dashboard.html" %}
 {% load i18n %}
 
-{% block title %}{% trans "Create project based on" %} {{ view.blueprint.title }} &mdash; {{ block.super }}{% endblock %}
+{% block title %}{% blocktrans with blueprint_title=view.blueprint.title %}Create {{ blueprint_title }}{% endblocktrans %} &mdash; {{ block.super }}{% endblock %}
 
 {% block dashboard_content %}
     <div class="l-menu">
         <div class="l-menu__content l-menu__content--aside">
-            <h1 class="u-first-heading">{% trans "Create project based on" %} {{ view.blueprint.title }}</h1>
+            <h1 class="u-first-heading">{% blocktrans with blueprint_title=view.blueprint.title %}Create {{ blueprint_title }}{% endblocktrans %}</h1>
 
             {% for error in form.non_field_errors %}
                 <span>{{ error }}</span>
@@ -26,7 +26,7 @@
 
                 <div class="u-spacer-bottom">
                     <input type="submit" class="btn btn--primary" name="send" value="{% trans 'Send' %}"/>
-                    <a href="{% url 'a4dashboard:blueprint-list' organisation_slug=view.organisation.slug %}"
+                    <a href="{% url 'a4dashboard:project-list' organisation_slug=view.organisation.slug %}"
                        class="btn btn--light">{% trans 'Cancel' %}</a>
                 </div>
             </form>
